requirements.txt
discord.py
requests
python-dotenv
fuzzywuzzy
python-Levenshtein  # Improves fuzzywuzzy performance
# openai